pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "gpt.5-chatter"
version = "0.1.0"
description = "GPT3.5 Chatter (Canton Of St. Gallen challenge)"
authors = [
    "Mudassar Zahid <mudassar.zahid@studium.uni-hamburg.de>",
    "Shreyasvi Natraj <shreyasvi.natraj@unige.ch>",
    "Justus Beck <justus.beck@campus.lmu.de>",
    "Mantas Kandratavičius <ge25ban@tum.de>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.0"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic = "^2.6.4"
requests = "^2.31.0"
httpx = "^0.27.0"
#torch = "^2.2.1"
openai = "^1.14.2"
python-dotenv = "^1.0.1"
elevenlabs = "1.0.0b1"
filelock = "3.13.1"
langchain = "^0.1.13"
chromadb = "^0.4.24"
langchain-openai = "^0.1.0"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
whisper = "^1.1.10"
openai-whisper = "^20231117"
langsmith = "^0.1.31"
pyht = "^0.0.26"
pydub = "^0.25.1"
streamlit = "^1.32.2"
pyaudio = "^0.2.14"
progressbar2 = "^4.4.2"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.31.0.20240311"
black = "^24.3.0"
mypy = "^1.9.0"
pydocstyle = "^6.3.0"
isort = "^5.13.2"

# Linting
[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pydocstyle]
convention = "google"
add-ignore = "D100,D104"

[tool.mypy]
python_version = "3.10"
warn_redundant_casts = "True"
warn_unused_ignores = "True"
warn_return_any = "True"
warn_unused_configs = "True"
warn_unreachable = "True"